[#22] Add support for hp-max formula
Can now set and configure the hp-max formula for monsters.
When "use formula" is enabled, the max hp value is not automatically set by the forge.
Clicking the "roll-hp" button on the artifact will roll and set the current/max hp values.
This allows GMM monsters to be used with token modules (such as Token Mold).